Downgrade to mybatis-spring 2.0.7.
Since at the time of the release only snapshot releases of mybatis-spring 2.1 are available we decided to downgrade to 2.0.7. for the release.

Unfortunately that release is not compatible with the current milestone of Spring Framework and therefore the MyBatis integration is broken.

If you want to use MyBatis with this release, manually add a dependency to mybatis-spring 2.1.0-SNAPSHOT.

Closes #1245