Further Refactored AggregatePath to offer more semantic result types TableInfo and ColumnInfo. They are now both records.
ColumnInfo might also gain a `type` element.
getIdColumnName() and getEffectiveIdColumnName() might also become ColumnInfo.

For the other methods I currently can't offer an improvement.

PersistentPropertyPathExtension is no longer part of the AggregatePath interface.

The methods navigating the aggregate path now use the streaming api instead of recursion.
This allowed to push helper methods down to the implementation, so they are no longer part of the interface.